regulator system is wrong size?
An attempt was made to construct a regulator system
(Control Law, Full Order) however the sizes do not match.
The system (Gz) has one input and one output however the
feedback (K), as produced using Ackermann's Formula
has three inputs for the 3rd order system.
Is this a case where the other values must be estimated?