all args to lsim1, regulator still isn't right
lsim1 has been expanded to support all the arguments.

The regulator system still doesn't work correctly and it
hasn't been determined why.  It needs an offset since
the regulator goes towards zero but the rpm should stabilize
at 700 rpm.  And the zeros chosen for K have a large impact
on the result.  These should influence how good the result
is, not whether it works at all.