got a Control Law example working, by intuition...
Difficulties have been encountered getting a Control Law
example working due to its particular requirements and
the characteristics of the engine model.  Specifically
the control input should reflect the output and it should
stabilize at zero for negative feedback to work.

An example has been designed intuitively.  Simply by making
the correct compensation to G(z) and choosing some K to
be feed back.  It is not the fastest, but it works and it
is stable.